refactor a Actividades
- se mueve a nuevo paquete actividades
- se creará aqui el tipo de actividad
