+++ src/main/java/com/asisge/consultifybackend/actividades/dominio/modelo/EstadoActividad.java
@@ -1,4 +1,4 @@
-package com.asisge.consultifybackend.proyectos.dominio.modelo;
+package com.asisge.consultifybackend.actividades.dominio.modelo;
 
 import lombok.Getter;
