+++ src/main/java/com/asisge/consultifybackend/actividades/dominio/modelo/Seguimiento.java
@@ -1,4 +1,4 @@
-package com.asisge.consultifybackend.proyectos.dominio.modelo;
+package com.asisge.consultifybackend.actividades.dominio.modelo;
 
 import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
 import lombok.AllArgsConstructor;
