+++ src/main/java/com/asisge/consultifybackend/actividades/infraestructura/controlador/ControladorSeguimiento.java
@@ -1,9 +1,9 @@
-package com.asisge.consultifybackend.proyectos.infraestructura.controlador;
+package com.asisge.consultifybackend.actividades.infraestructura.controlador;
 
 import com.asisge.consultifybackend.autenticacion.aplicacion.servicio.ServicioAutenticacion;
-import com.asisge.consultifybackend.proyectos.aplicacion.dto.NuevoSeguimientoDto;
-import com.asisge.consultifybackend.proyectos.aplicacion.servicio.ServicioSeguimientoActividad;
-import com.asisge.consultifybackend.proyectos.dominio.modelo.Seguimiento;
+import com.asisge.consultifybackend.actividades.aplicacion.dto.NuevoSeguimientoDto;
+import com.asisge.consultifybackend.actividades.aplicacion.servicio.ServicioSeguimientoActividad;
+import com.asisge.consultifybackend.actividades.dominio.modelo.Seguimiento;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
