+++ src/main/java/com/asisge/consultifybackend/proyectos/aplicacion/manejador/ManejadorNotificadorProyecto.java
@@ -3,10 +3,10 @@ package com.asisge.consultifybackend.proyectos.aplicacion.manejador;
 import com.asisge.consultifybackend.notificaciones.aplicacion.servicio.ServicioNotificacion;
 import com.asisge.consultifybackend.notificaciones.dominio.modelo.Notificacion;
 import com.asisge.consultifybackend.notificaciones.dominio.modelo.TipoNotificacion;
-import com.asisge.consultifybackend.proyectos.aplicacion.dto.ActividadDto;
-import com.asisge.consultifybackend.proyectos.aplicacion.servicio.NotificadorActividad;
+import com.asisge.consultifybackend.actividades.aplicacion.dto.ActividadDto;
+import com.asisge.consultifybackend.actividades.aplicacion.servicio.NotificadorActividad;
 import com.asisge.consultifybackend.proyectos.aplicacion.servicio.NotificadorProyecto;
-import com.asisge.consultifybackend.proyectos.dominio.modelo.Actividad;
+import com.asisge.consultifybackend.actividades.dominio.modelo.Actividad;
 import com.asisge.consultifybackend.proyectos.dominio.modelo.Proyecto;
 import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
 import com.asisge.consultifybackend.usuarios.dominio.modelo.UsuarioAutenticado;
