+++ src/main/java/com/asisge/consultifybackend/proyectos/aplicacion/manejador/ManejadorServicioTablero.java
@@ -3,9 +3,9 @@ package com.asisge.consultifybackend.proyectos.aplicacion.manejador;
 import com.asisge.consultifybackend.proyectos.aplicacion.dto.MiembroDto;
 import com.asisge.consultifybackend.proyectos.aplicacion.dto.TableroProyecto;
 import com.asisge.consultifybackend.proyectos.aplicacion.servicio.ServicioTablero;
-import com.asisge.consultifybackend.proyectos.dominio.modelo.Actividad;
+import com.asisge.consultifybackend.actividades.dominio.modelo.Actividad;
 import com.asisge.consultifybackend.proyectos.dominio.modelo.Proyecto;
-import com.asisge.consultifybackend.proyectos.dominio.puerto.RepositorioActividad;
+import com.asisge.consultifybackend.actividades.dominio.puerto.RepositorioActividad;
 import com.asisge.consultifybackend.proyectos.dominio.puerto.RepositorioProyecto;
 import com.asisge.consultifybackend.proyectos.dominio.puerto.RepositorioTablero;
 import com.asisge.consultifybackend.usuarios.dominio.modelo.Usuario;
